% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asSurvivalResult.R
\name{asSummarisedResult}
\alias{asSummarisedResult}
\title{Convert survival result back to summarised result}
\usage{
asSummarisedResult(result)
}
\arguments{
\item{result}{A survival_result object.}
}
\value{
A summarised_result object.
}
\description{
Internal function to convert a survival_result object back to a summarised_result format.
This is the inverse operation of asSurvivalResult().
}
\keyword{internal}
