% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockMGUS2cdm.R
\name{mockMGUS2cdm}
\alias{mockMGUS2cdm}
\title{Create mock CDM reference with survival::mgus2 dataset}
\usage{
mockMGUS2cdm()
}
\value{
CDM reference containing data from the survival::mgus2 dataset
}
\description{
Create mock CDM reference with survival::mgus2 dataset
}
\examples{
 \donttest{
cdm <- mockMGUS2cdm()
cdm$person
}
}
