% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSurvival.R
\name{optionsTableSurvival}
\alias{optionsTableSurvival}
\title{Additional arguments for the function tableSurvival()}
\usage{
optionsTableSurvival()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
tableSurvival and their given default value.
}
\examples{
{
optionsTableSurvival()
}


}
