% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cohortCount}
\alias{cohortCodelist}
\alias{settings}
\alias{attrition}
\alias{suppress}
\alias{bind}
\alias{exportSummarisedResult}
\alias{importSummarisedResult}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics:attrition]{attrition()}}, \code{\link[omopgenerics:bind]{bind()}}, \code{\link[omopgenerics:cohortCodelist]{cohortCodelist()}}, \code{\link[omopgenerics:cohortCount]{cohortCount()}}, \code{\link[omopgenerics:exportSummarisedResult]{exportSummarisedResult()}}, \code{\link[omopgenerics:importSummarisedResult]{importSummarisedResult()}}, \code{\link[omopgenerics:settings]{settings()}}, \code{\link[omopgenerics:suppress]{suppress()}}}
}}

