% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{survivalParticipants}
\alias{survivalParticipants}
\title{Participants contributing to a survival analysis}
\usage{
survivalParticipants(result)
}
\arguments{
\item{result}{Result object}
}
\value{
References to the study participants contributing to
a given analysis
}
\description{
Participants contributing to a survival analysis
}
\examples{
\donttest{
cdm <- mockMGUS2cdm()
surv <- estimateSingleEventSurvival(cdm,
                                    targetCohortTable = "mgus_diagnosis",
                                    outcomeCohortTable = "death_cohort",
                                    returnParticipants = TRUE)
survivalParticipants(surv)
}
}
