% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalSummary.R
\name{survivalSummary}
\alias{survivalSummary}
\title{Survival summary}
\usage{
survivalSummary(x)
}
\arguments{
\item{x}{Result from estimateSingleEventSurvival}
}
\value{
A tibble containing a summary of observed survival
}
\description{
Survival summary
}
\examples{
\donttest{
cdm <- mockMGUS2cdm()
surv <- estimateSingleEventSurvival(cdm,
                                    targetCohortTable = "mgus_diagnosis",
                                    outcomeCohortTable = "death_cohort")
survivalSummary(surv)
}
}
