% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCohortSurvival.R
\name{addCohortSurvival}
\alias{addCohortSurvival}
\title{Add survival information to a cohort table}
\usage{
addCohortSurvival(
  x,
  cdm,
  outcomeCohortTable,
  outcomeCohortId = 1,
  outcomeDateVariable = "cohort_start_date",
  censorOnCohortExit = FALSE,
  censorOnDate = NULL,
  followUpDays = Inf
)
}
\arguments{
\item{x}{cohort table to add survival information}

\item{cdm}{CDM reference}

\item{outcomeCohortTable}{The outcome cohort table of interest.}

\item{outcomeCohortId}{ID of event cohorts to include. Only one outcome
(and so one ID) can be considered.}

\item{outcomeDateVariable}{Variable containing date of outcome event}

\item{censorOnCohortExit}{If TRUE, an individual's follow up will be
censored at their cohort exit}

\item{censorOnDate}{if not NULL, an individual's follow up will be censored
at the given date}

\item{followUpDays}{Number of days to follow up individuals (lower bound 1,
upper bound Inf)}
}
\value{
Two additional columns will be added to x. The "time" column will
contain number of days to censoring. The "status" column will indicate
whether the patient had the event (value: 1), or did not have the event
(value: 0)
}
\description{
Add survival information to a cohort table
}
\examples{
cohort <- dplyr::tibble(
cohort_definition_id = c(1,1,1),
subject_id = c(1,2,3),
cohort_start_date = c(as.Date("2020-01-01"),
                      as.Date("2020-01-02"),
                      as.Date("2020-01-01")),
cohort_end_date = c(as.Date("2020-04-01"),
                    as.Date("2020-08-02"),
                    as.Date("2020-03-01"))
)
cdm <- PatientProfiles::mockPatientProfiles(cohort1 = cohort)
cdm$cohort1 <- cdm$cohort1 \%>\%
  addCohortSurvival(
    cdm = cdm,
    outcomeCohortTable = "cohort2",
    outcomeCohortId = 1
  )

}
