% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asSurvivalResult.R
\name{asSurvivalResult}
\alias{asSurvivalResult}
\title{A tidy implementation of the summarised_characteristics object.}
\usage{
asSurvivalResult(result)
}
\arguments{
\item{result}{A summarised_characteristics object.}
}
\value{
A tibble with a tidy version of the summarised_characteristics
object.
}
\description{
A tidy implementation of the summarised_characteristics object.
}
\examples{
\donttest{
cdm <- mockMGUS2cdm()
surv <- estimateSingleEventSurvival(
  cdm = cdm,
  targetCohortTable = "mgus_diagnosis",
  targetCohortId = 1,
  outcomeCohortTable = "death_cohort",
  outcomeCohortId = 1,
  eventGap = 7
) \%>\%
  asSurvivalResult()
}

}
