% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSurvival-package.R
\docType{package}
\name{CohortSurvival-package}
\alias{CohortSurvival}
\alias{CohortSurvival-package}
\title{CohortSurvival: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu.github.io/CohortSurvival/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
