% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurvival.R
\name{plotSurvival}
\alias{plotSurvival}
\title{Plot survival results}
\usage{
plotSurvival(
  result,
  x = "time",
  ylim = c(0, NA),
  ribbon = TRUE,
  facet = NULL,
  colour = NULL,
  colour_name = NULL
)
}
\arguments{
\item{result}{Survival results}

\item{x}{Variable to plot on x axis}

\item{ylim}{Limits for the Y axis}

\item{ribbon}{If TRUE, the plot will join points using a ribbon}

\item{facet}{Variables to use for facets}

\item{colour}{Variables to use for colours}

\item{colour_name}{Colour legend name}
}
\value{
A ggplot with the survival results plotted
}
\description{
Plot survival results
}
